pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jira-confluence-mcp"
version = "0.1.0"
description = "MCP server for Jira and Confluence Cloud with tiered safety gating"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.6.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
jira-confluence-mcp = "jira_confluence_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/jira_confluence_mcp"]

[tool.uv]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
